Chore(.global): separate security package
