+++ jacoco-exclude-class
@@ -1,6 +1,9 @@
 exclude com/newbarams/ajaja/AjajaApplication
 exclude com/newbarams/ajaja/global/config/*
 exclude com/newbarams/ajaja/global/common/*
+exclude com/newbarams/ajaja/global/security/*
+exclude com/newbarams/ajaja/global/security/jwt/*
+exclude com/newbarams/ajaja/global/security/filter/*
 exclude com/newbarams/ajaja/global/mock/*
 exclude com/newbarams/ajaja/infra/feign/common/*
 exclude com/newbarams/ajaja/infra/ses/*
